Исправил названия переменных и расположение функций
