+++ src/view/film-details.js
@@ -1,4 +1,8 @@
-import {createElement, showFullDate, isMultiple} from '../utils.js';
+import Genre from './genre.js';
+import Comment from './comment.js';
+import {createElement, showFullDate, isMultiple, RenderPosition, render} from '../utils.js';
+
+const body = document.body;
 
 const createFilmDetailsTemplate = (film) => {
   const {title, alternativeTitle, totalRating, poster, ageRating, director, writers, actors, runtime, genre, description} = film.filmInfo;
@@ -140,3 +144,32 @@ export default class FilmDetails {
     this._element = null;
   }
 }
+
+const renderFilmDetails = (film) => {
+  const filmDetailsComponent = new FilmDetails(film);
+
+  render(body, filmDetailsComponent.getElement(), RenderPosition.BEFOREEND);
+
+  const filmTableCells = filmDetailsComponent.getElement().querySelectorAll('.film-details__cell');
+  const commentsList = filmDetailsComponent.getElement().querySelector('.film-details__comments-list');
+
+  const {genre} = film.filmInfo;
+  const {comments} = film;
+
+  for (let i = 0; i < genre.length; i++) {
+    render(filmTableCells[filmTableCells.length - 1], new Genre(genre[i]).getElement(), RenderPosition.BEFOREEND);
+  }
+
+  for (let i = 0; i < comments.length; i++) {
+    render(commentsList, new Comment(comments[i]).getElement(), RenderPosition.BEFOREEND);
+  }
+
+  body.classList.add('hide-overflow');
+
+  filmDetailsComponent.getElement().querySelector('.film-details__close-btn').addEventListener('click', () => {
+    filmDetailsComponent.getElement().remove();
+    body.classList.remove('hide-overflow');
+  });
+};
+
+export {renderFilmDetails};
